src/BJMT.RsspII4net/SAI/TTS/Frames/SaiTtsFrameOffsetStart.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using BJMT.RsspII4net.Utilities;

namespace BJMT.RsspII4net.SAI.TTS.Frames
{
    class SaiTtsFrameOffsetStart : SaiTtsFrame
    {
        /// <summary>
        /// 获取/设置发起方周期
        /// </summary>
        public UInt32 SenderCycle { get; set; }


        public SaiTtsFrameOffsetStart()
            :base(SaiFrameType.TTS_OffsetStart, 0, 0, 0, 0)
        {
        }

        public SaiTtsFrameOffsetStart(ushort seqNo, uint senderTimestamp, uint senderCycle)
            :base(SaiFrameType.TTS_OffsetStart, seqNo, senderTimestamp, 0, 0)
        {
        }

        public override byte[] GetBytes()
        {
            var bytes = new byte[19];
            int startIndex = 0;

            // 消息类型
            bytes[startIndex++] = (byte)this.FrameType;

            // 序列号
            var tempBuf = RsspEncoding.ToNetUInt16(this.SequenceNo);
            Array.Copy(tempBuf, 0, bytes, startIndex, tempBuf.Length);
            startIndex += 2;

            // sender timestamp
            tempBuf = RsspEncoding.ToNetUInt32(this.SenderTimestamp);
            Array.Copy(tempBuf, 0, bytes, startIndex, tempBuf.Length);
            startIndex += 4;

            // Responsor last timestamp
            tempBuf = RsspEncoding.ToNetUInt32(this.ReceiverLastSendTimestamp);
            Array.Copy(tempBuf, 0, bytes, startIndex, tempBuf.Length);
            startIndex += 4;

            // sender last timestamp
            tempBuf = RsspEncoding.ToNetUInt32(this.SenderLastRecvTimestamp);
            Array.Copy(tempBuf, 0, bytes, startIndex, tempBuf.Length);
            startIndex += 4;

            // sender cycle
            tempBuf = RsspEncoding.ToNetUInt32(this.SenderCycle);
            Array.Copy(tempBuf, 0, bytes, startIndex, tempBuf.Length);
            startIndex += 4;

            return bytes;
        }

        public override void ParseBytes(byte[] bytes)
        {
            int startIndex = 0;

            // 消息类型
            this.FrameType = (SaiFrameType)bytes[startIndex++];

            // 序列号
            this.SequenceNo = RsspEncoding.ToHostUInt16(bytes, startIndex);
            startIndex += 2;

            // sender timestamp
            this.SenderTimestamp = RsspEncoding.ToHostUInt32(bytes, startIndex);
            startIndex += 4;

            // Responsor last timestamp
            this.ReceiverLastSendTimestamp = RsspEncoding.ToHostUInt32(bytes, startIndex);
            startIndex += 4;

            // sender last timestamp
            this.SenderLastRecvTimestamp = RsspEncoding.ToHostUInt32(bytes, startIndex);
            startIndex += 4;

            // sender cycle
            this.SenderCycle = RsspEncoding.ToHostUInt32(bytes, startIndex);
            startIndex += 4; 
        }
    }
}
